refactor: retire unreachable SAM3 node adapters
