+++ tests/fixtures/comfy_host_compatibility.v1.json
@@ -14,10 +14,10 @@
     ]
   },
   "expected_counts": {
-    "production_consumer_slots": 23,
-    "unique_production_modules": 17,
+    "production_consumer_slots": 21,
+    "unique_production_modules": 15,
     "root_monkeypatch_test_files": 0,
-    "canonical_monkeypatch_test_files": 3
+    "canonical_monkeypatch_test_files": 2
   },
   "external_evidence": {
     "github_public_code_search_2026_07_23": {
@@ -49,23 +49,12 @@
           "module": "easyuse_anima.aio.generation_normalization",
           "binder": null,
           "root_observation": "call_time"
-        },
-        {
-          "module": "easyuse_anima.nodes.impact_detailer_nodes",
-          "binder": null,
-          "root_observation": "call_time"
-        },
-        {
-          "module": "easyuse_anima.nodes.sam3_nodes",
-          "binder": null,
-          "root_observation": "call_time"
         }
       ],
       "repository_monkeypatch_consumers": {
         "root": [],
         "canonical": [
-          "tests/test_node_contracts.py",
-          "tests/test_sam3_nodes.py"
+          "tests/test_node_contracts.py"
         ]
       },
       "confirmed_external_consumers": [],
