+++ tests/test_python_package_skeleton.py
@@ -90,11 +90,9 @@ PACKAGE_MODULES = (
     "easyuse_anima.nodes",
     "easyuse_anima.nodes.aio_nodes",
     "easyuse_anima.nodes.input_types",
-    "easyuse_anima.nodes.impact_detailer_nodes",
     "easyuse_anima.nodes.prompt_advanced_nodes",
     "easyuse_anima.nodes.prompt_data_nodes",
     "easyuse_anima.nodes.regional_nodes",
-    "easyuse_anima.nodes.sam3_nodes",
     "easyuse_anima.profiles",
     "easyuse_anima.profiles.aio",
     "easyuse_anima.profiles.contract",
