+++ tests/test_registry_scanner_safety.py
@@ -73,13 +73,11 @@ EXPECTED_PYTHON_PACKAGE_FILES = {
     "easyuse_anima/nodes/__init__.py",
     "easyuse_anima/nodes/aio_nodes.py",
     "easyuse_anima/nodes/image_nodes.py",
-    "easyuse_anima/nodes/impact_detailer_nodes.py",
     "easyuse_anima/nodes/input_types.py",
     "easyuse_anima/nodes/lora_nodes.py",
     "easyuse_anima/nodes/naia_nodes.py",
     "easyuse_anima/nodes/prompt_data_nodes.py",
     "easyuse_anima/nodes/prompt_nodes.py",
-    "easyuse_anima/nodes/sam3_nodes.py",
     "easyuse_anima/nodes/wildcard_nodes.py",
     "easyuse_anima/profiles/__init__.py",
     "easyuse_anima/profiles/aio.py",
@@ -225,12 +223,10 @@ class RegistryScannerSafetyTests(unittest.TestCase):
             "easyuse_anima/naia/resolution.py",
             "easyuse_anima/nodes/aio_nodes.py",
             "easyuse_anima/nodes/image_nodes.py",
-            "easyuse_anima/nodes/impact_detailer_nodes.py",
             "easyuse_anima/nodes/lora_nodes.py",
             "easyuse_anima/nodes/naia_nodes.py",
             "easyuse_anima/nodes/prompt_data_nodes.py",
             "easyuse_anima/nodes/prompt_nodes.py",
-            "easyuse_anima/nodes/sam3_nodes.py",
             "easyuse_anima/nodes/wildcard_nodes.py",
             "easyuse_anima/registration.py",
             "easyuse_anima/prompt/correction.py",
